Move lerp functions used by UI view code into a common utility module.
